move header and footer into components folder
