+++ slimer37.github.io/src/App.tsx
@@ -1,5 +1,5 @@
-import Header from './Header.tsx'
-import Footer from './Footer.tsx'
+import Header from './components/Header.tsx'
+import Footer from './components/Footer.tsx'
 import NotFound from './NotFound.tsx'
 
 import './App.css'
